Support for mono to mono (#188)
* Added support for Mono to Mono channel layout.
Currently this only uses the L channel from the AW plugin and simply discards the R.
It needs to be considered if a L+R/2 is a better solution, which it probably is.

* Updated Perl script to deduce which plugins are mono only using the MacAU version.
AirwinRegistry now has a isMono attribute.
Tried to update the Editor to only show mono compatible plugins when running in pure mono mode, but not sure I got them all...

* Override current AWProcessor when plugin is loaded as M->M if it does not support mono
Currently it's overridden with "Chamber" which is the only mono reverb from the recommended list

* Added option to "force" certain plugins to be usable in mono mode, as for instance ToTape8.
As I understood Chris there is only slight variations in the flutter in that plugin so it should be
okay to use in mono (I know I want to :-)

* Updated the Jog method to take mono plugins into consideration